refactor(storage): reorganize the directory structure
