+++ storage/src/memory/mod.rs
@@ -1,0 +1,14 @@
+//! In-memory storage implementations
+//!
+//! These implementations use DashMap for lock-free concurrent access.
+//! Suitable for testing and single-node deployments without persistence.
+
+pub mod event_store;
+pub mod anchor_store;
+pub mod cf_store;
+pub mod object_store;
+
+pub use event_store::EventStore;
+pub use anchor_store::AnchorStore;
+pub use cf_store::CFStore;
+pub use object_store::MemoryObjectStore;
