+++ storage/src/rocks/core/error.rs
